+++ src/main/java/com/habit/server/controller/TaskController.java
@@ -11,15 +11,19 @@ import java.io.OutputStream;
  */
 public class TaskController {
     private final DatabaseTeamManager teamManager;
-
+ 
     public TaskController(DatabaseTeamManager teamManager) {
         this.teamManager = teamManager;
     }
 
+    public HttpHandler getSaveTaskHandler() {
+        return new SaveTaskHandler();
+    }
+ 
     public HttpHandler getAddTaskHandler() {
         return new AddTaskHandler();
     }
-
+ 
     public HttpHandler getGetTasksHandler() {
         return new GetTasksHandler();
     }
@@ -195,6 +199,55 @@ public class TaskController {
             java.io.OutputStream os = exchange.getResponseBody();
             os.write(response.getBytes("UTF-8"));
             os.close();
+        }
+    }
+    // --- タスク保存API ---
+    class SaveTaskHandler implements com.sun.net.httpserver.HttpHandler {
+        @Override
+        public void handle(com.sun.net.httpserver.HttpExchange exchange) throws java.io.IOException {
+            if (!"POST".equals(exchange.getRequestMethod())) {
+                String response = "POSTメソッドのみ対応";
+                exchange.sendResponseHeaders(405, response.getBytes().length);
+                try (java.io.OutputStream os = exchange.getResponseBody()) {
+                    os.write(response.getBytes());
+                }
+                return;
+            }
+            byte[] bodyBytes = exchange.getRequestBody().readAllBytes();
+            String bodyStr = (bodyBytes != null) ? new String(bodyBytes, java.nio.charset.StandardCharsets.UTF_8) : "";
+            String response;
+            try {
+                // key=value&...形式で受信
+                String[] params = bodyStr.split("&");
+                java.util.Map<String, String> map = new java.util.HashMap<>();
+                for (String param : params) {
+                    String[] kv = param.split("=", 2);
+                    if (kv.length < 2) continue;
+                    map.put(kv[0], java.net.URLDecoder.decode(kv[1], "UTF-8"));
+                }
+                String taskId = map.get("taskId");
+                String taskName = map.get("taskName");
+                String description = map.getOrDefault("description", "");
+                int estimatedMinutes = Integer.parseInt(map.getOrDefault("estimatedMinutes", "0"));
+                boolean isTeamTask = Boolean.parseBoolean(map.getOrDefault("isTeamTask", "false"));
+                String dueTimeStr = map.get("dueTime");
+                java.time.LocalTime dueTime = dueTimeStr != null && !dueTimeStr.isEmpty() ? java.time.LocalTime.parse(dueTimeStr) : null;
+                String cycleType = map.getOrDefault("cycleType", "daily");
+                String teamID = map.get("teamID");
+                // repeatDaysは未対応
+                com.habit.domain.Task task = new com.habit.domain.Task(
+                    taskId, taskName, description, estimatedMinutes,
+                    java.util.Collections.emptyList(), isTeamTask, dueTime, cycleType
+                );
+                new com.habit.server.TaskRepository().saveTask(task, teamID);
+                response = "タスク保存成功";
+            } catch (Exception ex) {
+                response = "タスク保存失敗: " + ex.getMessage();
+            }
+            exchange.sendResponseHeaders(200, response.getBytes().length);
+            try (java.io.OutputStream os = exchange.getResponseBody()) {
+                os.write(response.getBytes());
+            }
         }
     }
 }
